requirements.txt
djangorestframework
python-decouple
Pillow
psycopg2-binary
djangorestframework-simplejwt
django-filter
pytelegrambotapi
